requirements.txt
browser-use==0.2.5
faiss-cpu==1.10.0
python-dotenv
langchain-ollama
asyncio 